added Menu folder and front end axios request
