+++ client/src/Menu/Sides.json
