--- client/src/utils/Api.js
@@ -1,11 +1,0 @@
-// import axios from "axios";
-
-// axios.get("/chicken", { params: { q: query } });
-
-// axios.get("/beef", { params: { q: query }});
-
-// axios.get("/sides", { params: {q: query }});
-
-// axios.get("/dessert", { params: {q: query }});
-
-// export default axios;
